requirements.txt
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
cx-Oracle==8.3.0
debugpy==1.6.3
decorator==5.1.1
Django==4.0.6
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap4==22.2
django-phone-field==1.8.1
django-phonenumber-field==6.3.0
django-tempus-dominus==5.1.2.17
django-widget-tweaks==1.4.12
entrypoints==0.4
executing==1.1.1
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.3.5
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
packaging==21.3
parso==0.8.3
pexpect==4.8.0
phonenumbers==8.12.52
pickleshare==0.7.5
Pillow==9.2.0
pkg_resources==0.0.0
prompt-toolkit==3.0.31
psutil==5.9.2
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pyzmq==24.0.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.5.1
tornado==6.2
traitlets==5.4.0
wcwidth==0.2.5
